fix(docker): update config.ini for more generalized development
Extra options are necessary for more generalized development. The verbose-http-errors allows allows for error assertions to be readable. The mongodb additions enable the capturing of ancillary block info necessary for capturing and using inline actions. Finally, removing the whitelist mongodb filter allows for developers to run and capture their own actions (while not excluding the capturing of the demo).