fix(docker): update config.ini for more generalized development (#86)
Extra options are necessary for more generalized development. The verbose-http-errors allows allows for error assertions to be readable. The mongodb additions enable the capturing of ancillary block info necessary for capturing and using inline actions. Finally, removing the whitelist mongodb filter allows for developers to run and capture their own actions (while not excluding the capturing of the demo).

If any of these disagree with the community's vision though, feel free to disregard, but figured I'd make a PR for this since I find this to be the default setup I keep incorporating myself.

The main consideration is the removal of the whitelist. I know you have a commented out wildcard that one could simply toggle on, but it seemed like an extra unnecessary step in my mind.